Move time related helper to seperate module(#201)
Closes: #201
